+++ tests/integration/react.js
@@ -28,7 +28,7 @@ const App = () => (
 
   it('jsx 2', () => {
     const input = `
-const App = () => (<div><Component attribute1="value1" attribute2="value2" attribute3={{a:1,b:2,c:3}}><span>foo</span></Component></div>);
+const App = () => (<div><Component attribute1="value1" attribute2="value2" attribute3={{a:1,b:2,c:3}}>foo</Component></div>);
 `;
     const output = `
 const App = () => (
@@ -38,7 +38,7 @@ const App = () => (
       attribute2="value2"
       attribute3={{ a: 1, b: 2, c: 3 }}
     >
-      <span>foo</span>
+      foo
     </Component>
   </div>
 );